feat: move ApiResponseSchema to a new file and update import paths
